Handle WP NEM12 edge-case with missing MSATSLoadDateTime col
Western Power NEM12 files don't always include an empty column for MSATSLoadDateTime if it is never present, instead the column can be completely absent.
This occurs with files downloaded from their internal portal and causes a ValueError exception to be unnecessarily raised on parse, as the number of columns is less than expected.

The AEMO "Meter Data File Format Specification NEM12 & NEM13" states that MSATSLoadDateTime is required if available, but is not mandatory.
As such, instead of erroring when the final MSATSLoadDateTime column isn't present, we can proceed and fill it with None with no consequences.
Validation for parse_300_row has been modified to ensure that the error is only raised if number of required non-reading fields is less than 6, instead of equals 7.
These WP NEM12 files otherwise seem to parse fine with this change.